Prevent words overlapping & fill blank cells
Changed to prevent words overlapping if the letters on the point of
overlapping are different. Added code to limit number of attempts to
place a word with no such overlaps, as this may fail after many attempts
and get stuck in a seemingly infinite loop. Also added code to fill
remaining blank cells in the wordsearch array with random letters.