Changed so words help block also appears initially
Changed so that the help block for the words input that describes how
words should be entered is displayed to begin with (before the user
enters anything into the field) as well as when the input is invalid
after the user has entered something into the field. Stored maximum
number of words into a variable which is used in the regex expression
and in the help block. Also renamed "constants" to use naming convention
for constants (capital letters with underscores separating words).